Fixes the key value being left on the stack when using iterators (the other case that's not "and we're not actually using this for iteration", asides from the empty table case). Closes #111 - fixed Closes #110 - need to test exactly how much extra speed was gained Closes #108 - seems to be fixed, albeit std::mutex is a butt on VC++

+++ sol/debug.hpp
@@ -35,7 +35,7 @@ inline std::string dump_types(lua_State* L) {
         if(i != 1) {
             visual += " | ";
         }
-        visual += type_name(L, stack::get<type>(L, i));
+        visual += type_name(L, stack::get<type>(L, static_cast<int>(i)));
     }
     return visual;
 }
